tests: extract duplicated _make_stub_timeout into tests/stub_helpers.py
Six test files each defined a byte-identical no-op timeout bash-stub
builder (differing only in docstring wording). Consolidate into a
single owner, matching the existing single-owner precedent
(tests/ec2_stub.py, tests/launcher_blocks.py), so future changes to
the stub's semantics can't silently desync across call sites.

test_ec2_launcher_kill.py now imports _make_stub_timeout directly from
tests/stub_helpers.py instead of re-exported via test_ec2_fetch_branch.
