+++ tests/test_ec2_seed_auth.py
@@ -23,6 +23,7 @@ import os
 import subprocess
 from pathlib import Path
 
+from tests.stub_helpers import _make_stub_timeout
 from tests.test_ec2_transport import _stub_timeout as _make_killing_stub_timeout
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
@@ -154,25 +155,6 @@ esac
 """
     )
     stub_path.chmod(0o755)
-
-
-def _make_stub_timeout(stub_dir: Path) -> None:
-    """No-op `timeout`: run the child, ignore the cap.
-
-    Adequate for tests that merely need the binary to exist on the
-    stubbed PATH. A test that asserts the cap actually *fires* must
-    use `_make_killing_stub_timeout` (imported from
-    test_ec2_transport) instead.
-    """
-    stub = stub_dir / "timeout"
-    stub.write_text(
-        """#!/usr/bin/env bash
-while [[ "$1" == --* ]]; do shift; done
-shift
-exec "$@"
-"""
-    )
-    stub.chmod(0o755)
 
 
 def _make_stub_git(stub_path: Path, *, name: str = "Test User",
