+++ tests/test_seed_inspect_dirs.py
@@ -38,6 +38,8 @@ import os
 import subprocess
 from pathlib import Path
 
+from tests.stub_helpers import _make_stub_timeout
+
 REPO_ROOT = Path(__file__).resolve().parent.parent
 SEED_SH = REPO_ROOT / "scripts" / "remote" / "seed-repo.sh"
 
@@ -177,20 +179,6 @@ esac
 """
     )
     stub_path.chmod(0o755)
-
-
-def _make_stub_timeout(stub_dir: Path) -> None:
-    """Same `timeout` stub as test_seed_repo_sh.py — macOS doesn't ship
-    GNU `timeout` in /usr/bin."""
-    stub = stub_dir / "timeout"
-    stub.write_text(
-        """#!/usr/bin/env bash
-while [[ "$1" == --* ]]; do shift; done
-shift
-exec "$@"
-"""
-    )
-    stub.chmod(0o755)
 
 
 def _make_git_repo(root: Path, name: str, files: dict[str, str],
